+++ src-docs/src/views/icon/icons.js
@@ -71,6 +71,7 @@ export const iconTypes = [
   'download',
   'email',
   'empty',
+  'eql',
   'exit',
   'expand',
   'expandMini',